Shrink installation

+++ install.sh
@@ -1,27 +1,27 @@
 #!/usr/bin/sh
 
-# backup folder
-mkdir -v ~/dotfiles/backup
+read -p "Are you sure? This will overwrite existing dotfiles... y/n" -n 1 -r
+echo    # (optional) move to a new line
+if [[ ! $REPLY =~ ^[Yy]$ ]]
+then
+    [[ "$0" = "$BASH_SOURCE" ]] && exit 1 || return 1 # handle exits from shell or function but don't exit interactive shell
+fi
 
 # vim
 echo "########## vim"
 ln -sf ~/dotfiles/vim ~/.vim
 ln -sf ~/dotfiles/vimrc ~/.vimrc
-
 # neovim
 echo "########## nvim"
 ln -sf ~/dotfiles/vim ~/.config/nvim
 ln -sf ~/dotfiles/vimrc ~/.config/nvim/init.vim
-
 # vim pathogen
 echo "########## pathogen"
 mkdir ~/dotfiles/vim/bundle
 mkdir ~/dotfiles/vim/autoload
-
 # download pathogen to temp file ~, activate o
 rm -v "~/dotfiles/vim/autoload/pathogen.vim"
 wget -O ~/dotfiles/vim/autoload/pathogen.vim https://tpo.pe/pathogen.vim
-
 # vim plugins
 echo "########## neocomplete"
 git clone https://github.com/Shougo/neocomplete.vim.git ~/dotfiles/vim/bundle/neocomplete.vim
@@ -36,23 +36,10 @@ git clone https://github.com/ntpeters/vim-better-whitespace.git ~/dotfiles/vim/b
 echo "########## vim go"
 git clone https://github.com/fatih/vim-go.git ~/dotfiles/vim/bundle/vim-go
 echo "########## update all vim plugins"
-cd ~/dotfiles/vim/bundle
-for i in `ls`; do
-  cd "$i"
-  git pull
-  cd ..
-done
+for i in ~/dotfiles/vim/bundle/*; do git -C $i pull; done
 
 # terminator
 echo "########## terminator"
-if [ -L ~/.config/terminator ]
-then
-    rm ~/.config/terminator
-elif [ -e ~/.config/terminator ]
-then
-	echo "backup existing terminator config"
-	mv ~/.config/terminator ~/dotfiles/backup
-fi
 ln -sf ~/dotfiles/terminator ~/.config/terminator
 
 # sorin zprezto
